test: split heap snapshot limit tests
test/parallel/test-heapsnapshot-near-heap-limit.js is timing out in CI
on low-memory and slow-CPU devices. Split off the worker test to its own
test file to allow the test to finish in time.

Refs: https://github.com/nodejs/node/issues/36961
PR-URL: https://github.com/nodejs/node/pull/37189
